renaming to lmdb_m
