+++ setup.py
@@ -105,7 +105,7 @@ if patch_lmdb_source:
     except Exception:
         pass
     shutil.copytree('lib', dest)
-
+    breakpoint()
     # Copy away the lmdb source then patch it
     if sys.platform.startswith('win'):
         patchfile = 'lib' + os.sep + 'py-lmdb' + os.sep + 'env-copy-txn.patch'
@@ -149,7 +149,7 @@ if sys.platform.startswith('win'):
 # configuration may differ, forcing a recompile (and therefore likely compile
 # errors). This happens even when `use_cpython` since user might want to
 # LMDB_FORCE_CFFI=1 during testing.
-with open('lmdb/_config.py', 'w') as fp:
+with open('lmdb_m/_config.py', 'w') as fp:
     fp.write('CONFIG = dict(%r)\n\n' % ((
         ('extra_compile_args', extra_compile_args),
         ('extra_sources', extra_sources),
@@ -167,7 +167,7 @@ if use_cpython:
                                '-I' + os.path.dirname(memsink.__file__)]
     ext_modules = [Extension(
         name='cpython',
-        sources=['lmdb/cpython.c'] + extra_sources,
+        sources=['lmdb_m/cpython.c'] + extra_sources,
         extra_compile_args=extra_compile_args,
         libraries=libraries,
         include_dirs=extra_include_dirs,
@@ -177,21 +177,21 @@ else:
     print('Using cffi extension.')
     install_requires = ['cffi>=0.8']
     try:
-        import lmdb.cffi
-        ext_modules = [lmdb.cffi._ffi.verifier.get_extension()]
+        import lmdb_m.cffi
+        ext_modules = [lmdb_m.cffi._ffi.verifier.get_extension()]
     except ImportError:
         sys.stderr.write('Could not import lmdb; ensure cffi is installed!\n')
         ext_modules = []
 
 def grep_version():
-    path = os.path.join(os.path.dirname(__file__), 'lmdb/__init__.py')
+    path = os.path.join(os.path.dirname(__file__), 'lmdb_m/__init__.py')
     with open(path) as fp:
         for line in fp:
             if line.startswith('__version__'):
                 return eval(line.split()[-1])
 
 setup(
-    name='lmdb',
+    name='lmdb_m',
     version=grep_version(),
     description="Universal Python binding for the LMDB 'Lightning' Database",
     long_description="Universal Python binding for the LMDB 'Lightning' Database",
@@ -200,8 +200,8 @@ setup(
     maintainer='Nic Watson',
     license='OLDAP-2.8',
     url='http://github.com/jnwatson/py-lmdb/',
-    packages=['lmdb'],
-
+    packages=['lmdb_m'],
+    python_requires=">=3.12",
     classifiers=[
         "Programming Language :: Python",
         "Programming Language :: Python :: Implementation :: CPython",
@@ -219,7 +219,7 @@ setup(
         "Topic :: Database",
         "Topic :: Database :: Database Engines/Servers",
     ],
-    ext_package='lmdb',
+    ext_package='lmdb_m',
     ext_modules=ext_modules,
     install_requires=install_requires,
 )
